+++ src/main/java/com/fiap/tech/application/controllers/cliente/identifica/requests/IdentificaClienteRequest.java
@@ -1,0 +1,4 @@
+package com.fiap.tech.application.controllers.cliente.identifica.requests;
+
+public record IdentificaClienteRequest(String nome, String cpf, String email) {
+}
