--- src/main/java/com/fiap/tech/application/controllers/cliente/requests/IdentificaClienteRequest.java
@@ -1,4 +1,0 @@
-package com.fiap.tech.application.controllers.cliente.requests;
-
-public record IdentificaClienteRequest(String nome, String cpf, String email) {
-}
