--- src/main/java/com/fiap/tech/application/controllers/produto/store/requests/StoreProdutoRequest.java
@@ -1,15 +1,0 @@
-package com.fiap.tech.application.controllers.produto.store.requests;
-
-import com.fiap.tech.domain.enums.produto.CategoriaEnum;
-
-import java.util.Date;
-
-public record StoreProdutoRequest(
-        String nome,
-        Float valor,
-        String descricao,
-        CategoriaEnum categoria,
-        Integer quantidade,
-        Date dataCriacao
-) {
-}
