--- src/main/java/com/fiap/tech/application/controllers/produto/update/requests/UpdateProdutoRequest.java
@@ -1,14 +1,0 @@
-package com.fiap.tech.application.controllers.produto.update.requests;
-
-import com.fiap.tech.domain.enums.produto.CategoriaEnum;
-
-import java.util.Date;
-
-public record UpdateProdutoRequest(
-        String nome,
-        Float valor,
-        String descricao,
-        CategoriaEnum categoria,
-        Integer quantidade,
-        Date dataCriacao) {
-}
